requirements.txt
pyaes==1.6.1
pyasn1==0.5.0
pyfiglet==0.8.post1
rsa==4.9
Telethon==1.29.3
python-dotenv==1.0.0
